+++ CHANGELOG.md
@@ -13,6 +13,8 @@ Version 1.2.2 (not released yet)
 - optimize Flamethower rendering (client-side)
 - add 'debug' command for enabling RF built-in debugging features
 - fix 'Orion bug' - default 'miner1' entity spawning periodically on clients (server-side)
+- fix beeping every frame if chat input buffer is full
+- apply proper chat input limits client-side so PF server does not kick the sender
 - fix multiple security vulnerabilities
 - stability improvements